pyproject.toml
[tool.poetry]
name = "rag-tutorial"
version = "0.1.0"
description = ""
authors = ["otomoryosuke <xwtfw614@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.12,<3.13"
python-dotenv = "^1.0.1"
chromadb = "0.5.2"
langchain = "0.2.5"
langchain-cli = "0.0.25"
langchain-openai = "0.1.8"
langchain-community = "0.2.5"
langchain-chroma = "0.1.1"
langchainhub = "0.1.20"
streamlit = "1.35.0"
pyyaml = "^6.0.2"
beautifulsoup4 = "^4.12.3"
logging = "^0.4.9.6"
google-cloud-kms = "^3.0.0"
pybase64 = "^1.4.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
